polish and refactor
