+++ apps/io-sign-backoffice-app/src/app/[locale]/[institution]/api-keys/[api-key]/_components/ApiKeyProvider.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import { ApiKey } from "@/lib/api-keys";
+import { ApiKeyContext } from "@/lib/api-keys/client";
+
+export type Props = {
+  value: ApiKey;
+  children: React.ReactNode;
+};
+
+export default function ApiKeyProvider(props: Props) {
+  return <ApiKeyContext.Provider {...props} />;
+}
