+++ apps/io-sign-backoffice-app/src/app/api/api-keys/route.ts
@@ -1,10 +1,11 @@
 import { NextResponse, NextRequest } from "next/server";
 
+import { createApiKeyPayloadSchema } from "@/lib/api-keys";
+
 import {
-  createApiKeyPayloadSchema,
   createApiKey,
   ApiKeyAlreadyExistsError,
-} from "@/lib/api-keys";
+} from "@/lib/api-keys/use-cases";
 
 import { ZodError } from "zod";
