+++ apps/io-sign-backoffice-app/src/app/api/institutions/[institution]/api-keys/[api-key]/route.ts
@@ -2,11 +2,8 @@ import { NextRequest, NextResponse } from "next/server";
 
 import { z, ZodError } from "zod";
 
-import {
-  cidrSchema,
-  fiscalCodeSchema,
-  upsertApiKeyField,
-} from "@/lib/api-keys";
+import { cidrSchema, fiscalCodeSchema } from "@/lib/api-keys";
+import { upsertApiKeyField } from "@/lib/api-keys/use-cases";
 
 const pathSchema = z.object({
   institution: z.string().uuid(),
