+++ apps/io-sign-backoffice-app/src/lib/api-keys/apim.ts
@@ -1,3 +1,5 @@
+import "server-only";
+
 import { getApimClient } from "../apim";
 
 import { ApiKey, ApiKeyWithSecret } from "./index";
