+++ apps/io-sign-backoffice-app/src/lib/api-keys/client.ts
@@ -1,0 +1,7 @@
+"use client";
+
+import { createContext } from "react";
+
+import { ApiKey } from "./index";
+
+export const ApiKeyContext = createContext<ApiKey | undefined>(undefined);
