+++ apps/io-sign-backoffice-app/src/lib/api-keys/cosmos.ts
@@ -1,3 +1,5 @@
+import "server-only";
+
 import { getCosmosConfig, getCosmosContainerClient } from "@/lib/cosmos";
 import { FeedResponse } from "@azure/cosmos";
