+++ apps/io-sign-backoffice-app/src/lib/api-keys/use-cases.ts
@@ -1,0 +1,84 @@
+import "server-only";
+
+import { apiKeyExists, insertApiKey, getApiKeys, getApiKey } from "./cosmos";
+
+import {
+  createApiKeySubscription,
+  deleteApiKeySubscription,
+  exposeApiKeySecret,
+} from "./apim";
+
+import { ApiKey, ApiKeyWithSecret, CreateApiKeyPayload } from "./index";
+import { ulid } from "ulid";
+
+function ApiKey(payload: CreateApiKeyPayload): ApiKey {
+  const apiKey = {
+    ...payload,
+    id: ulid(),
+    createdAt: new Date(),
+    status: "active" as const,
+  };
+  return apiKey;
+}
+
+export class ApiKeyAlreadyExistsError extends Error {
+  constructor(cause = {}) {
+    super("the API key already exists");
+    this.name = "ApiKeyAlreadyExistsError";
+    this.cause = cause;
+  }
+}
+
+export async function createApiKey(payload: CreateApiKeyPayload) {
+  try {
+    // check if the api key for the given input already exists
+    const apiKeyAlreadyExists = await apiKeyExists(
+      payload.institutionId,
+      payload.displayName
+    );
+    if (apiKeyAlreadyExists) {
+      throw new ApiKeyAlreadyExistsError(
+        "such name already exists for the institution"
+      );
+    }
+    const apiKey = ApiKey(payload);
+    await createApiKeySubscription(apiKey);
+    try {
+      await insertApiKey(apiKey);
+    } catch (e) {
+      await deleteApiKeySubscription(apiKey);
+      throw new Error("unable to create the API key", { cause: e });
+    }
+    return apiKey.id;
+  } catch (e) {
+    throw e instanceof ApiKeyAlreadyExistsError
+      ? e
+      : new Error("unable to create the API key", { cause: e });
+  }
+}
+
+export async function* listApiKeys(institutionId: string) {
+  try {
+    const apiKeys = getApiKeys(institutionId);
+    for await (const apiKey of apiKeys) {
+      yield exposeApiKeySecret(apiKey);
+    }
+  } catch (e) {
+    throw new Error("unable to get the API keys", { cause: e });
+  }
+}
+
+export async function getApiKeyWithSecret(
+  id: string,
+  institutionId: string
+): Promise<ApiKeyWithSecret> {
+  try {
+    const apiKey = await getApiKey(id, institutionId);
+    return exposeApiKeySecret(apiKey);
+  } catch (e) {
+    console.log(e);
+    throw new Error("unable to get the API Key", { cause: e });
+  }
+}
+
+export { upsertApiKeyField } from "./cosmos";
